feat(llm): finalize vision context and memory workflow
