+++ ballontranslator/utils/llm_profiles.py
@@ -16,6 +16,7 @@ from typing import (
     get_type_hints,
 )
 
+from ballontranslator.utils.logger import logger as LOGGER
 from ballontranslator.utils.secret_store import SecretStore
 from ballontranslator.utils.structures import Config, field, nested_dataclass
 
@@ -25,7 +26,17 @@ LLM_OCR_KEY = "LLMOCR"
 LLM_INPAINT_KEY = "LLMInpaint"
 OLD_LLM_TRANSLATORS = ("ChatGPT", "ChatGPT_exp", "LLM_API_Translator")
 
-THINKING_LEVEL_OPTIONS = ["None", "minimal", "low", "medium", "high", "xhigh"]
+THINKING_AUTO = "Auto"
+THINKING_DISABLED = "Disabled"
+THINKING_LEVEL_OPTIONS = [
+    THINKING_AUTO,
+    THINKING_DISABLED,
+    "minimal",
+    "low",
+    "medium",
+    "high",
+    "xhigh",
+]
 VISION_DETAIL_LEVEL_OPTIONS = ["None", "auto", "low", "high"]
 PROVIDER_ALIASES = {
     "Google": "Gemini",
@@ -184,7 +195,7 @@ class LLMProfile(Config):
     image_base_url: str = ""
     image_model: str = ""
     image_model_options: List[str] = field(default_factory=list)
-    thinking_level: str = "None"
+    thinking_level: str = THINKING_AUTO
     thinking_level_options: List[str] = field(default_factory=lambda: list(THINKING_LEVEL_OPTIONS))
     prompt: str = DEFAULT_TRANSLATION_PROMPT
     vision_prompt: str = DEFAULT_OCR_PROMPT
@@ -207,12 +218,78 @@ class LLMProfile(Config):
         return copy.deepcopy(self.__dict__)
 
 
+def normalize_thinking_level(value: Any) -> str:
+    """Return the canonical persisted reasoning-control value.
+
+    >>> normalize_thinking_level('None')
+    'Auto'
+    >>> normalize_thinking_level('disable')
+    'Disabled'
+    """
+    if not isinstance(value, str):
+        return THINKING_AUTO
+    level = value.strip()
+    if level.lower() in {'', 'none', 'auto'}:
+        return THINKING_AUTO
+    if level.lower() in {'disable', 'disabled'}:
+        return THINKING_DISABLED
+    return level
+
+
+def _normalize_profile_thinking(profile: LLMProfile) -> LLMProfile:
+    """Migrate legacy reasoning settings without dropping custom options.
+
+    >>> profile = LLMProfile(
+    ...     thinking_level='None',
+    ...     thinking_level_options=['None', 'high'],
+    ... )
+    >>> migrated = _normalize_profile_thinking(profile)
+    >>> migrated.thinking_level, migrated.thinking_level_options
+    ('Auto', ['Auto', 'Disabled', 'high'])
+    """
+    raw_level = profile.thinking_level
+    if not isinstance(raw_level, str):
+        LOGGER.warning(
+            'Discard invalid LLM profile thinking_level for %s.',
+            profile.id or profile.name or '<unnamed>',
+        )
+    profile.thinking_level = normalize_thinking_level(raw_level)
+
+    raw_options = profile.thinking_level_options
+    if not isinstance(raw_options, list):
+        LOGGER.warning(
+            'Discard invalid LLM profile thinking_level_options for %s.',
+            profile.id or profile.name or '<unnamed>',
+        )
+        raw_options = THINKING_LEVEL_OPTIONS
+    elif any(not isinstance(option, str) for option in raw_options):
+        LOGGER.warning(
+            'Discard invalid entries from LLM profile '
+            'thinking_level_options for %s.',
+            profile.id or profile.name or '<unnamed>',
+        )
+
+    normalized_options = []
+    for option in [THINKING_AUTO, THINKING_DISABLED, *raw_options]:
+        if not isinstance(option, str):
+            continue
+        option = normalize_thinking_level(option)
+        if option not in normalized_options:
+            normalized_options.append(option)
+    if profile.thinking_level not in normalized_options:
+        normalized_options.append(profile.thinking_level)
+    profile.thinking_level_options = normalized_options
+    return profile
+
+
 def profile_from_config(profile: Any) -> LLMProfile:
     if isinstance(profile, LLMProfile):
-        return copy.deepcopy(profile)
-    if isinstance(profile, Mapping):
-        return LLMProfile(**copy.deepcopy(dict(profile)))
-    raise TypeError(f"Unsupported LLM profile config: {type(profile)!r}")
+        loaded = copy.deepcopy(profile)
+    elif isinstance(profile, Mapping):
+        loaded = LLMProfile(**copy.deepcopy(dict(profile)))
+    else:
+        raise TypeError(f"Unsupported LLM profile config: {type(profile)!r}")
+    return _normalize_profile_thinking(loaded)
 
 
 def profile_to_dict(profile: Any) -> Dict:
@@ -293,7 +370,7 @@ def profiles_from_json(value: str) -> List[LLMProfile]:
         data.pop('profile_type', None)
         data = _normalize_import_profile_data(data)
         try:
-            profiles.append(LLMProfile(**data))
+            profiles.append(profile_from_config(data))
         except (TypeError, ValueError):
             continue
     return profiles
@@ -518,7 +595,7 @@ def _infer_provider(provider: str, base_url: str, model: str) -> str:
 def normalize_model(provider: str, model: str) -> Tuple[str, str]:
     provider = canonical_provider(provider)
     model = _strip_model_prefix(model)
-    thinking = "None"
+    thinking = THINKING_AUTO
     if provider == "OpenAI":
         aliases = {
             "gpt3": "gpt-5.5",
@@ -535,7 +612,7 @@ def normalize_model(provider: str, model: str) -> Tuple[str, str]:
             thinking = "high"
         elif model == "deepseek-chat":
             model = "deepseek-v4-flash"
-            thinking = "None"
+            thinking = THINKING_AUTO
     if not model:
         model = PROVIDER_DEFAULTS[provider]["model"]
     return model, thinking
@@ -576,7 +653,7 @@ def profile_from_old_settings(old_key: str, params: Dict, secret_store: SecretSt
     if provider in {"OpenAI", "DeepSeek"} and not using_override:
         if model not in PROVIDER_DEFAULTS[provider]["model_options"]:
             model = PROVIDER_DEFAULTS[provider]["model"]
-            thinking = "None"
+            thinking = THINKING_AUTO
     api_key = _old_api_key(old_key, params)
     require_key = PROVIDER_DEFAULTS[provider]["require_api_key"]
     if require_key and not api_key:
